Allow omitting `mode` argument
The builtin `open` can be called w/o a mode argument, but `MockOpen`
default mode was an explicit `'r'`. Since in a test context it makes
more sense to test the actual call and not the "expected" call, i.e.:
```python
def test_case():
    with path('builtins.open', new_callable=MockOpen) as mock_open:
        with open('/some/file') as f:
            pass

    assert f.called_once_with('/some/file') # This is how mocks are usually asserted.
    # assert f.called_once_with('/some/file', 'r') # This isn't intuitive.
```

We now honor the actual call made when registering the mock call,
without the default `'r'` mode argument if not provided.